refactor(react-ink): move useTextBuffer beside TextInput
