+++ .changeset/textbuffer-relocation.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-ink": patch
+---
+
+refactor: move useTextBuffer beside TextInput
