+++ packages/react-ink/src/primitives/textInput/useTextBuffer.ts
